Only improve "Prime Video" navigation bar for signed-in users
I mistakenly thought that the design of the navigation bar differed
between sites, so I created seperate styles for the two variants I
encountered. Now, I realized that the style only depends on whether
the user is signed in, so I removed the styles for the second (not
signed in) variant.