Instantly enable improved navigation on movie/series pages
Because the `<html>` element on these pages always has the ID
`dv-product-details`, the JavaScript check is not needed there and a CSS
ID selector can be used to apply the styles instantly without "jumping".
On all other "Prime Video" pages, the JavaScript check will still be
used until I find a similar solution for them.